==> 1_preprocess.ipynb <==
#!pip install wget
import wget, tarfile
import pandas as pd
import numpy as np
import pickle
import gc

wget.download("https://tesla-ap-shanghai-1256322946.cos.ap-shanghai.myqcloud.com/cephfs/tesla_common/deeplearning/dataset/algo_contest/train_preliminary.zip")
wget.download("https://tesla-ap-shanghai-1256322946.cos.ap-shanghai.myqcloud.com/cephfs/tesla_common/deeplearning/dataset/algo_contest/test.zip")
wget.download("https://tesla-ap-guangzhou-1256322946.cos.ap-guangzhou.myqcloud.com/cephfs/tesla_common/deeplearning/dataset/algo_contest/train_semi_final.zip")

!mkdir ./data
!unzip test.zip -d ./data
!unzip train_preliminary.zip -d ./data
!unzip train_semi_final.zip -d ./data

ad_test = pd.read_csv('./data/test/ad.csv')
click_test = pd.read_csv('./data/test/click_log.csv')
ad_train = pd.read_csv('./data/train_preliminary/ad.csv')
ad_semi_train = pd.read_csv('./data/train_semi_final/ad.csv')
click_train = pd.read_csv('./data/train_preliminary/click_log.csv')
click_semi_train = pd.read_csv('./data/train_semi_final/click_log.csv')

ad_all_train = pd.concat([ad_train, ad_semi_train], axis=0).reset_index(drop=True)
click_all_train = pd.concat([click_train, click_semi_train], axis=0).reset_index(drop=True)
ad_all_train = ad_all_train.drop_duplicates().reset_index(drop=True)
click_all_train = click_all_train.drop_duplicates().reset_index(drop=True)

del ad_train,ad_semi_train,click_train,click_semi_train
gc.collect()

train_data = click_all_train.merge(ad_all_train, 'left', 'creative_id')
test_data = click_test.merge(ad_test, 'left', 'creative_id')
all_data = pd.concat([train_data, test_data], axis=0, sort=False).sort_values(by=['time', 'user_id']).reset_index(drop=True)
train_data.to_pickle('./data/train.pkl')
test_data.to_pickle('./data/test.pkl')
all_data.to_pickle('./data/data.pkl')
