==> 3_w2v.ipynb <==
import os
import pickle
import gc
import gensim
import time
from gensim import corpora, models, similarities
from gensim.models.doc2vec import TaggedDocument
import pandas as pd
import numpy as np
from tqdm import tqdm
tqdm.pandas()

for col in ['creative_id', 'ad_id', 'product_id', 'product_category', 'advertiser_id', 'industry']:
    list_d = pd.read_pickle('./idlist/{}_list.pkl'.format(col))
    list_d = [list(str(item)) for item in list_d]
    model = models.Word2Vec(list_d,sg=1,min_count=1,size=128,window=10,workers=30,iter=10)
    We = []
    if '0' in model.wv:
        for i in tqdm(range(len(model.wv.index2word))):
            We.append(model.wv[str(i)].reshape((1,-1)))
    else:
        We.append(np.zeros((1, 128)))
        for i in tqdm(range(len(model.wv.index2word))):
            We.append(model.wv[str(i+1)].reshape((1,-1)))
    We = np.vstack(We)
    np.save('./w2v_128_10/{}_embedding_weight.npy'.format(col), We)
