==> 8_blending.ipynb <==
import pandas as pd
import numpy as np
import os
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import StratifiedKFold
from sklearn.metrics import accuracy_score
import warnings
warnings.filterwarnings('ignore')

result = pd.read_csv('./result.csv')
result.head()

res_test = []
for root, dirs, files in os.walk('./age_res/'):
    for files_name in files:
        if res_test == []:
            res_test = np.load(root+files_name)
        else:
            res_test += np.load(root+files_name)

result['predicted_age'] = np.argmax(res_test, axis=1) + 1

res_test = []
for root, dirs, files in os.walk('./gender_res/'):
    for files_name in files:
        if res_test == []:
            res_test = np.load(root+files_name)
        else:
            res_test += np.load(root+files_name)

result['predicted_gender'] = np.argmax(res_test, axis=1) + 1

result.to_csv('result_final_3.csv', index=False)
result.head()

from ti import session
ti_session = session.Session()
inputs = ti_session.upload_data(path="result_final_3.csv", bucket="algo-1258864734", key_prefix="upload_semi")

res = pd.read_csv('result_final_2.csv')

res.predicted_age.corr(result.predicted_age)
